some changes

+++ schemas/orderSchema.js
@@ -2,7 +2,10 @@ import z from "zod";
 import orderItemSchema from "./orderItemSchema.js";
 
 const orderSchema = z.object({
-  userId: z.number().min(1, "User ID must be a number and greater than 1"),
+  userId: z
+    .number()
+    .int()
+    .min(1, "User ID must be a number and greater than 1"),
   products: z.array(orderItemSchema).min(1),
 });
